requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
python-dotenv
httpx
pytest
pytest-asyncio
pytest-cov
python-multipart
